Address #384:Implement LABELFONT and SETLABELFONT.
LABELFONT
Outputs a list describing the current font that is used by the LABEL command.
This output is a list of the form
[[FaceName] Height unused unused Weight Italic unused unused unused unused unused unused unused]

SETLABELFONT font
Sets the font to be used by the LABEL command.
font can be a list or a word.
List:
[[FaceName] Height unused unused Weight Italic unused unused unused unused unused unused unused]
Word:
CSS font value, as in https://developer.mozilla.org/en-US/docs/Web/CSS/font