Implement MOUSEON and demo SCRIBBLE. MOUSEON leftbuttondown leftbuttonup rightbuttondown rightbuttonup move
Starts trapping mouse events that are sent to the canvas and passes
them to leftbuttondown, leftbuttonup, rightbuttondown, rightbuttonup,
or move, depending on the type of event. The callbacks may either be
a word that is the name of a procedure to call, or a list of instructions
to run.

Example:
MOUSEON [SETPOS MOUSEPOS PENDOWN] [PENUP] [] [] [SETPOS MOUSEPOS]

demo "SCRIBBLE
Move the turtle with mouse; left button down is to draw; right button
down is to erase.